Moving testAlertEnqueue as an integration test since it is now usign TSDB Service
